refactor/config test
