+++ .eslintignore
@@ -2,3 +2,4 @@ test/_fixtures
 test/datasource/**/_fixtures
 test/manager/**/_fixtures
 test/platform/**/_fixtures
+test/config/**/_fixtures
