+++ .prettierignore
@@ -13,4 +13,6 @@ test/_fixtures/
 test/datasource/**/_fixtures
 test/manager/**/_fixtures
 test/platform/**/_fixtures
+test/config/**/_fixtures
+
 bin/yarn*
