--- test/_fixtures/config/index.js
@@ -1,5 +1,0 @@
-const defaultConfig = require('../../../lib/config/defaults').getConfig();
-
-module.exports = {
-  ...defaultConfig,
-};
