+++ test/config/config/_fixtures/index.js
@@ -1,0 +1,5 @@
+const defaultConfig = require('../../../../lib/config/defaults').getConfig();
+
+module.exports = {
+  ...defaultConfig,
+};
