+++ test/config/decrypt.spec.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const { decryptConfig } = require('../../lib/config/decrypt.js');
 
-const privateKey = fs.readFileSync('test/_fixtures/keys/private.pem');
+const privateKey = fs.readFileSync('test/config/keys/_fixtures/private.pem');
 
 describe('config/decrypt', () => {
   describe('decryptConfig()', () => {
