+++ test/config/file.spec.js
@@ -1,6 +1,6 @@
 const path = require('path');
 const file = require('../../lib/config/file.js');
-const customConfig = require('../_fixtures/config/file');
+const customConfig = require('./config/_fixtures/file');
 
 describe('config/file', () => {
   describe('.getConfig()', () => {
@@ -10,16 +10,13 @@ describe('config/file', () => {
       );
     });
     it('parses custom config file', () => {
-      const configFile = path.resolve(__dirname, '../_fixtures/config/file.js');
+      const configFile = path.resolve(__dirname, './config/_fixtures/file.js');
       expect(file.getConfig({ RENOVATE_CONFIG_FILE: configFile })).toEqual(
         customConfig
       );
     });
     it('migrates', () => {
-      const configFile = path.resolve(
-        __dirname,
-        '../_fixtures/config/file2.js'
-      );
+      const configFile = path.resolve(__dirname, './config/_fixtures/file2.js');
       const res = file.getConfig({ RENOVATE_CONFIG_FILE: configFile });
       expect(res).toMatchSnapshot();
       expect(res.rangeStrategy).toEqual('bump');
