+++ test/config/npm/_fixtures/02.json
