+++ test/config/npm/_fixtures/coffeelint.json
