+++ test/config/npm/_fixtures/next.json
