+++ test/config/npm/_fixtures/plocktest1/package.json
