+++ test/config/npm/_fixtures/renovate-config-default.json
