+++ test/config/npm/_fixtures/renovate-config-group.json
