+++ test/config/npm/_fixtures/renovate-config-ikatyang.json
