+++ test/config/npm/_fixtures/renovate-config-monorepo.json
