+++ test/config/npm/_fixtures/renovate-config-packages.json
