+++ test/config/npm/_fixtures/typescript.json
