+++ test/config/npm/_fixtures/vue-test-utils.json
