+++ test/config/npm/_fixtures/vue.json
