+++ test/config/npm/_fixtures/webpack.json
