+++ test/workers/repository/error-config.spec.js
@@ -5,7 +5,7 @@ const {
 let config;
 beforeEach(() => {
   jest.resetAllMocks();
-  config = require('../../_fixtures/config');
+  config = require('../../config/config/_fixtures');
 });
 
 describe('workers/repository/error-config', () => {
