+++ test/workers/repository/error.spec.js
@@ -5,7 +5,7 @@ jest.mock('../../../lib/workers/repository/error-config');
 let config;
 beforeEach(() => {
   jest.resetAllMocks();
-  config = require('../../_fixtures/config');
+  config = require('../../config/config/_fixtures');
 });
 
 describe('workers/repository/error', () => {
