+++ test/workers/repository/init/vulnerability.spec.js
@@ -1,7 +1,9 @@
 let config;
 beforeEach(() => {
   jest.resetAllMocks();
-  config = JSON.parse(JSON.stringify(require('../../../_fixtures/config')));
+  config = JSON.parse(
+    JSON.stringify(require('../../../config/config/_fixtures'))
+  );
 });
 
 const {
