+++ test/workers/repository/process/limits.spec.js
@@ -4,7 +4,7 @@ const limits = require('../../../../lib/workers/repository/process/limits');
 let config;
 beforeEach(() => {
   jest.resetAllMocks();
-  config = { ...require('../../../_fixtures/config') };
+  config = { ...require('../../../config/config/_fixtures') };
 });
 
 describe('workers/repository/process/limits', () => {
