+++ test/workers/repository/process/write.spec.js
@@ -10,7 +10,7 @@ limits.getPrsRemaining = jest.fn(() => 99);
 let config;
 beforeEach(() => {
   jest.resetAllMocks();
-  config = { ...require('../../../_fixtures/config') };
+  config = { ...require('../../../config/config/_fixtures') };
 });
 
 describe('workers/repository/write', () => {
