+++ test/workers/repository/result.spec.js
@@ -3,7 +3,7 @@ const { processResult } = require('../../../lib/workers/repository/result');
 let config;
 beforeEach(() => {
   jest.resetAllMocks();
-  config = require('../../_fixtures/config');
+  config = require('../../config/config/_fixtures');
 });
 
 describe('workers/repository/result', () => {
