+++ test/workers/repository/updates/flatten.spec.js
@@ -5,7 +5,7 @@ const {
 let config;
 beforeEach(() => {
   jest.resetAllMocks();
-  config = { ...require('../../../_fixtures/config') };
+  config = { ...require('../../../config/config/_fixtures') };
   config.errors = [];
   config.warnings = [];
 });
